+++ source/Collection/CollectionView.js
@@ -75,6 +75,13 @@ export default class CollectionView extends Component {
     scrollLeft: PropTypes.number,
 
     /**
+     * Controls scroll-to-cell behavior of the Grid.
+     * The default ("auto") scrolls the least amount possible to ensure that the specified cell is fully visible.
+     * Use "start" to align cells to the top/left of the Grid and "end" to align bottom/right.
+     */
+    scrollToAlignment: PropTypes.oneOf(['auto', 'end', 'start']).isRequired,
+
+    /**
      * Cell index to ensure visible (by forcefully scrolling if necessary).
      */
     scrollToCell: PropTypes.number,
@@ -100,6 +107,7 @@ export default class CollectionView extends Component {
     noContentRenderer: () => null,
     onScroll: () => null,
     onSectionRendered: () => null,
+    scrollToAlignment: 'auto',
     style: {}
   };
 
@@ -139,7 +147,13 @@ export default class CollectionView extends Component {
   componentDidMount () {
     const { cellLayoutManager, scrollLeft, scrollToCell, scrollTop } = this.props
 
-    this._scrollbarSize = getScrollbarSize()
+    // If this component was first rendered server-side, scrollbar size will be undefined.
+    // In that event we need to remeasure.
+    if (!this._scrollbarSizeMeasured) {
+      this._scrollbarSize = getScrollbarSize()
+      this._scrollbarSizeMeasured = true
+      this.setState({})
+    }
 
     if (scrollToCell >= 0) {
       this._updateScrollPositionForScrollToCell()
@@ -166,7 +180,7 @@ export default class CollectionView extends Component {
 
   componentDidUpdate (prevProps, prevState) {
     const { height, scrollToCell, width } = this.props
-    const { scrollLeft, scrollPositionChangeReason, scrollTop } = this.state
+    const { scrollLeft, scrollPositionChangeReason, scrollToAlignment, scrollTop } = this.state
 
     // Make sure requested changes to :scrollLeft or :scrollTop get applied.
     // Assigning to scrollLeft/scrollTop tells the browser to interrupt any running scroll animations,
@@ -193,6 +207,7 @@ export default class CollectionView extends Component {
     // Update scroll offsets if the current :scrollToCell values requires it
     if (
       height !== prevProps.height ||
+      scrollToAlignment !== prevProps.scrollToAlignment ||
       scrollToCell !== prevProps.scrollToCell ||
       width !== prevProps.width
     ) {
@@ -207,6 +222,16 @@ export default class CollectionView extends Component {
     const { cellLayoutManager } = this.props
 
     cellLayoutManager.calculateSizeAndPositionData()
+
+    // If this component is being rendered server-side, getScrollbarSize() will return undefined.
+    // We handle this case in componentDidMount()
+    this._scrollbarSize = getScrollbarSize()
+    if (this._scrollbarSize === undefined) {
+      this._scrollbarSizeMeasured = false
+      this._scrollbarSize = 0
+    } else {
+      this._scrollbarSizeMeasured = true
+    }
   }
 
   componentWillUnmount () {
@@ -302,11 +327,13 @@ export default class CollectionView extends Component {
     // Force browser to hide scrollbars when we know they aren't necessary.
     // Otherwise once scrollbars appear they may not disappear again.
     // For more info see issue #116
-    if (totalHeight <= height) {
-      collectionStyle.overflowY = 'hidden'
-    }
-    if (totalWidth <= width) {
+    const verticalScrollBarSize = totalHeight > height ? this._scrollbarSize : 0
+    const horizontalScrollBarSize = totalWidth > width ? this._scrollbarSize : 0
+    if (totalWidth + verticalScrollBarSize <= width) {
       collectionStyle.overflowX = 'hidden'
+    }
+    if (totalHeight + horizontalScrollBarSize <= height) {
+      collectionStyle.overflowY = 'hidden'
     }
 
     return (
@@ -440,11 +467,12 @@ export default class CollectionView extends Component {
   }
 
   _updateScrollPositionForScrollToCell () {
-    const { cellLayoutManager, height, scrollToCell, width } = this.props
+    const { cellLayoutManager, height, scrollToAlignment, scrollToCell, width } = this.props
     const { scrollLeft, scrollTop } = this.state
 
     if (scrollToCell >= 0) {
       const scrollPosition = cellLayoutManager.getScrollPositionForCell({
+        align: scrollToAlignment,
         cellIndex: scrollToCell,
         height,
         scrollLeft,