+++ source/Grid/Grid.test.js
@@ -1,7 +1,8 @@
+import getScrollbarSize from 'dom-helpers/util/scrollbarSize'
 import React from 'react'
 import { findDOMNode } from 'react-dom'
-import { render } from '../TestUtils'
 import { Simulate } from 'react-addons-test-utils'
+import { render } from '../TestUtils'
 import Grid from './Grid'
 
 const NUM_ROWS = 100
@@ -43,6 +44,7 @@ describe('Grid', () => {
     rowHeight = 20,
     rowCount = NUM_ROWS,
     scrollLeft,
+    scrollToAlignment,
     scrollToColumn,
     scrollToRow,
     scrollTop,
@@ -67,6 +69,7 @@ describe('Grid', () => {
         rowHeight={rowHeight}
         rowCount={rowCount}
         scrollLeft={scrollLeft}
+        scrollToAlignment={scrollToAlignment}
         scrollToColumn={scrollToColumn}
         scrollToRow={scrollToRow}
         scrollTop={scrollTop}
@@ -117,23 +120,75 @@ describe('Grid', () => {
   })
 
   describe('shows and hides scrollbars based on rendered content', () => {
-    it('should set overflowX:hidden on scroll-container if columns fit within the available width', () => {
-      const rendered = findDOMNode(render(getMarkup({ columnCount: 4 })))
+    let scrollbarSize
+
+    beforeAll(() => {
+      scrollbarSize = getScrollbarSize()
+    })
+
+    it('should set overflowX:hidden if columns fit within the available width and y-axis has no scrollbar', () => {
+      const rendered = findDOMNode(render(getMarkup({
+        columnCount: 4,
+        rowCount: 5
+      })))
       expect(rendered.style.overflowX).toEqual('hidden')
     })
 
-    it('should leave overflowX:auto on scroll-container if columns require more than the available width', () => {
-      const rendered = findDOMNode(render(getMarkup({ columnCount: 25 })))
+    it('should set overflowX:hidden if columns and y-axis scrollbar fit within the available width', () => {
+      const rendered = findDOMNode(render(getMarkup({
+        columnCount: 4,
+        width: 200 + scrollbarSize
+      })))
+      expect(rendered.style.overflowX).toEqual('hidden')
+    })
+
+    it('should leave overflowX:auto if columns require more than the available width', () => {
+      const rendered = findDOMNode(render(getMarkup({
+        columnCount: 4,
+        width: 200 - 1,
+        rowCount: 5
+      })))
       expect(rendered.style.overflowX).not.toEqual('hidden')
     })
 
-    it('should set overflowY:hidden on scroll-container if rows fit within the available height', () => {
-      const rendered = findDOMNode(render(getMarkup({ rowCount: 5 })))
+    it('should leave overflowX:auto if columns and y-axis scrollbar require more than the available width', () => {
+      const rendered = findDOMNode(render(getMarkup({
+        columnCount: 4,
+        width: 200 + scrollbarSize - 1
+      })))
+      expect(rendered.style.overflowX).not.toEqual('hidden')
+    })
+
+    it('should set overflowY:hidden if rows fit within the available width and xaxis has no scrollbar', () => {
+      const rendered = findDOMNode(render(getMarkup({
+        rowCount: 5,
+        columnCount: 4
+      })))
       expect(rendered.style.overflowY).toEqual('hidden')
     })
 
-    it('should leave overflowY:auto on scroll-container if rows require more than the available height', () => {
-      const rendered = findDOMNode(render(getMarkup({ rowCount: 25 })))
+    it('should set overflowY:hidden if rows and x-axis scrollbar fit within the available width', () => {
+      const rendered = findDOMNode(render(getMarkup({
+        rowCount: 5,
+        height: 100 + scrollbarSize
+      })))
+      expect(rendered.style.overflowY).toEqual('hidden')
+    })
+
+    it('should leave overflowY:auto if rows require more than the available width', () => {
+      const rendered = findDOMNode(render(getMarkup({
+        rowCount: 5,
+        height: 100 - 1,
+        columnCount: 4
+      })))
+      expect(rendered.style.overflowY).not.toEqual('hidden')
+    })
+
+    it('should leave overflowY:auto if rows and x-axis scrollbar require more than the available width', () => {
+      const rendered = findDOMNode(render(getMarkup({
+        rowCount: 5,
+        height: 100 + scrollbarSize - 1
+      })))
       expect(rendered.style.overflowY).not.toEqual('hidden')
     })
   })
@@ -209,6 +264,40 @@ describe('Grid', () => {
       }))
       expect(grid.state.scrollLeft).toEqual(2350)
       expect(grid.state.scrollTop).toEqual(1920)
+    })
+
+    it('should scroll to the correct position for :scrollToAlignment "start"', () => {
+      const grid = render(getMarkup({
+        scrollToAlignment: 'start',
+        scrollToColumn: 24,
+        scrollToRow: 49
+      }))
+      // 100 columns * 50 item width = 5,000 total item width
+      // 100 rows * 20 item height = 2,000 total item height
+      // 4 columns and 5 rows can be visible at a time.
+      // The minimum amount of scrolling leaves the specified cell in the bottom/right corner (just scrolled into view).
+      // Since alignment is set to "start" we should scroll past this point until the cell is aligned top/left.
+      expect(grid.state.scrollLeft).toEqual(1200)
+      expect(grid.state.scrollTop).toEqual(980)
+    })
+
+    it('should scroll to the correct position for :scrollToAlignment "end"', () => {
+      render(getMarkup({
+        scrollToColumn: 99,
+        scrollToRow: 99
+      }))
+      const grid = render(getMarkup({
+        scrollToAlignment: 'end',
+        scrollToColumn: 24,
+        scrollToRow: 49
+      }))
+      // 100 columns * 50 item width = 5,000 total item width
+      // 100 rows * 20 item height = 2,000 total item height
+      // We first scroll past the specified cell and then back.
+      // The minimum amount of scrolling then should leave the specified cell in the top/left corner (just scrolled into view).
+      // Since alignment is set to "end" we should scroll past this point until the cell is aligned bottom/right.
+      expect(grid.state.scrollLeft).toEqual(1050)
+      expect(grid.state.scrollTop).toEqual(900)
     })
   })